usb: gadget: Fix incorrect ISOC data transfer PIDs
The PIDs for Iscohronous data transfers are incorrect
for high bandwidth endpoints when the request length is
less than the wMaxPacketSize.
The correct PIDs for ISOC data transfers is
For request length < maxPayloadSize
	- DATA0,
For maxPayloadSize < length < 2*maxPayloadSize
	- DATA0,DATA1
For 2*maxPayloadSize <  length < 3*maxPayloadSize
	- DATA2, DATA1, DATA0.
Fix this by setting the value of trb->size depending
on request length rather than fixing it to the value
depending on wMaxPacketSize.

Change-Id: I9ae88ed3abc566e02daa727e84a2938582931436
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>