ANDROID: cpufreq: uid_concurrent_active_time
In order to model the energy used by the cpu,
we need to expose cpu time information to
userspace. Output the time each uid spent running
with 1 to num_possible_cpus online and not idle.

The time should also include the time spent by
pids that are no longer alive.

We can use this information to blame pids and
uids for the energy they are using.

For instance, if 4 cores are online and active
for 50ms and the uid is running on one of the
cores, the uid should be blamed for 1/4 of the
base energy used by the cpu when 1 or more cores is
active. Using other proc files, the uid can
also be blamed for the cost of keeping a cluster
on and the cost of the individual cores it
is using.

b/68399339

Test: cat /proc/uid_concurrent_active_time

Change-Id: Idba8d86ad3e1650102638b0da6308941e70cd36e
Signed-off-by: Marissa Wall <marissaw@google.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>