Revert "power: qpnp-smbcharger: Report charger types in REAL_TYPE property"
During the merge of LA.UM.6.6.r1-06500-89xx.0, the changes to
drivers/power/qpnp-smbcharger.c were not applied because the
areas that were modified by this patch were not present or had
been changed by OnePlus during the dash charging bring up.

Unfortunately, this appears to have broken the device being
recognized by the computer for MTP or adb because the REAL_TYPE
property is not getting set.

This reverts commit 9dbf8291e20493ff380440ed96406ed3ff065ab1.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>