modules: don't unload wlan.ko
I don't know what OnePlus (or Qcom?) is doing, and this has been tested on a
100% stock, factory reset, unmodified build of OxygenOS 5.0.8, but when the device
screens off, and it doesn't need to be in deep sleep suspend mode, uevent sends a
signal and WiFi crumbles down. wpa_supplicant service is killed and the WiFi userspace
stack notifies the Kernel and it disconnects WiFi from inside the qcacld drivers and
then unloads the module, no questions asked. I can't seem to figure out a way to
totally block the entire disconnection because I can't differentiate between a user
WiFi shutdown or a shutdown by userspace. The only thing I can seem to do is prevent
the module from being unloaded, which somewhat removes a sizeable delay when the
entire thing is trying to get back on its feet when the screen turns on.
It's still a shit solution, but it's what I can do to make things less fucked.

And yes I also tested this with a custom wlan.ko from caf from the latest tag,
it's still fucked to high heavens since the issue clearly comes from fucking
userspace.

To whoever implemented this WiFi teardown on screen off (also this is regardless of
what the user has choosen as the WiFi sleep policy in Android), fuck you.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>